MNT align Brant with model conventions
